refactor findMatchedSignature to be shared in livy
